BLIP2/requirements.txt
altair==5.1.2
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bleach==6.1.0
blinker==1.6.3
blis==0.7.11
braceexpand==0.1.7
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.3.0
click==8.1.7
cloudpathlib==0.15.1
confection==0.1.3
contexttimer==0.3.3
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
decorator==5.1.1
decord==0.6.0
diffusers==0.16.0
distlib==0.3.7
einops==0.7.0
executing==2.0.0
fairscale==0.4.4
filelock==3.12.4
fonttools==4.43.1
fsspec==2023.9.2
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.37
huggingface-hub==0.17.3
identify==2.5.30
idna==3.4
imageio==2.31.5
importlib-metadata==6.8.0
importlib-resources==6.1.0
iopath==0.1.10
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kaggle==1.5.16
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.1
nodeenv==1.8.0
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
packaging==23.2
pandas==2.0.3
parso==0.8.3
pathy==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
plotly==5.17.0
portalocker==2.8.2
pre-commit==3.5.0
preshed==3.0.9
prompt-toolkit==3.0.39
protobuf==4.24.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pycocoevalcap==1.2
pycocotools==2.0.7
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-magic==0.4.27
python-slugify==8.0.1
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
safetensors==0.4.0
-e git+https://github.com/salesforce/LAVIS.git@7f00a0891b2890843f61c002a8e9532a40343648#egg=salesforce_lavis
scikit-image==0.21.0
scipy==1.10.1
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
spacy==3.7.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
streamlit==1.27.2
sympy==1.12
tenacity==8.2.3
text-unidecode==1.3
thinc==8.2.1
tifffile==2023.7.10
timm==0.4.12
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==1.12.1+cu102
torchaudio==0.12.1+cu102
torchvision==0.13.1+cu102
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
transformers==4.25.0
triton==2.1.0
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.6
validators==0.22.0
virtualenv==20.24.5
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.8
weasel==0.3.2
webdataset==0.2.57
webencodings==0.5.1
zipp==3.17.0
